# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RentMitra API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get public user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUser"

  # ── LISTINGS ──────────────────────────────────────────────────────────────
...
  /recently-viewed:
    get:
      operationId: getRecentlyViewed
      tags: [recently-viewed]
      summary: Get current user's recently viewed listings
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recently viewed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"

    post:
      operationId: trackView
      tags: [recently-viewed]
      summary: Track a listing view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackViewInput"
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── BUSINESS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      required: [id, name, email, userType, isVerified, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        profilePhoto:
          type: ["string", "null"]
        userType:
          type: string
          enum: [individual, business, admin]
        isVerified:
          type: boolean
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PublicUser:
      type: object
      required: [id, name, userType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        profilePhoto:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        userType:
          type: string
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        listingCount:
          type: integer
        businessProfile:
          oneOf:
            - $ref: "#/components/schemas/BusinessProfile"
            - type: "null"
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
        city:
          type: string
        state:
          type: string
...
    RentalPrice:
      type: object
      properties:
        daily:
          type: ["number", "null"]
        weekly:
          type: ["number", "null"]
        monthly:
          type: ["number", "null"]
    Listing:
      type: object
      required: [id, ownerId, title, category, status, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        brand:
          type: ["string", "null"]
        condition:
          type: string
          enum: [new, like_new, good, fair, poor]
        rentalPrice:
          $ref: "#/components/schemas/RentalPrice"
        city:
          type: string
        state:
          type: string
        pincode:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        thumbnails:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        isFeatured:
          type: boolean
        rejectionReason:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        owner:
          oneOf:
            - $ref: "#/components/schemas/PublicUser"
            - type: "null"
...
    TrackViewInput:
      type: object
      required: [listingId]
      properties:
        listingId:
          type: integer

    # ── BUSINESS ───────────────────────────────────────────────────────────
    BusinessProfile:
      type: object
      required: [id, userId, businessName]
      properties:
        id:
          type: integer
        userId:
          type: integer
        businessName:
          type: string
        logo:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        gstNumber:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...